Use a splat `*` expression to output the private IP addresses of all aws_instances.
The splat expression captures all objects in a list that share an attribute. The special * symbol iterates over all the elements of a given list and returns information based on the shared attribute you define.
Without the splat expression, Terraform would not be able to output the entire array of your instances and would only return the first item in the array.